requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langgraph>=0.0.10
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
web3>=6.0.0
fpdf>=1.7.2
pydantic>=2.0.0
typing-extensions>=4.5.0
python-multipart>=0.0.6
redis>=5.0.0
numpy>=1.24.0
tiktoken>=0.5.0
aiohttp>=3.9.0
requests>=2.31.0

# Development tools
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0 